requirements.txt
streamlit 
pytube 
python-dotenv
whisper 
gemini_pro